move data transformers to adapters directory, add files from desktop
